vectorized function calls with automated masking
Given an Enoki array with pointers to instances of a type T, Enoki can
now transparently perform vector calls to T's methods -- for example:

    Array<T*, N> values = <...>;
    values->my_function();

Only a single function call is needed when the all elements of the array
have the same value. Otherwise, Enoki performs a sequence of masked
function calls. Arguments and return values are automatically propagated
if provided.

Note that some instrumentation using the ENOKI_CALL_HELPER_* macros is
needed for this to work -- see the 'tests/calls.cpp' example for
details.