redesigned masks to be top-level primitives
Previously masks classes were hidden away in detail namespace, and they
could not use the CRTP approach for customization. This commit makes
them a top-level primitive analogous to enoki::Array. The possibility to
customize mask behavior is used to introduce matched Packet and
PacketMask arrays that always broadcast to the trailing dimension.