arithmetic involving arrays of pointers
This commit extends Enoki with the ability to add/subtract an integer
to/from an Enoki pointer array. The operation has the same semantics as
in C: specifically, the integer amount is scaled by the size of the
underlying type.

Taking differences between two pointers arrays returns an integer array,
whose entries represent the pointer difference divided by the size of
the underlying type.