Fix CMAKE_CUDA_FLAGS (#32)
When using a double hyphen to specify the cudart flag for CUDA, I get the following error in CMakeTestCUDACompiler:

    nvcc fatal   : redefinition of argument 'cudart'

This seems to be because the tests adds "-cudart' if it does not already sees it in flags, hence leading to a command line with both "--cudart" and "-cudart". I hope there is not the inverse behavior on some other setups.