LLDB introspection script for Enoki array types
When debugging programs built on top of Enoki using LLDB, the stringified
versions of arrays  are needlessly verbose and reveal private
implementation details. For instance, printing a simple 3D vectors like
Array<float, 3>(1, 2, 3) yields

    (enoki::Array<float, 3>) $0 = {
      enoki::StaticArrayImpl<float, 3, false, enoki::Array<float, 3>, int> = {
        enoki::StaticArrayImpl<float, 4, false, enoki::Array<float, 3>, int> = {
          m = (1, 2, 3, 0)
        }
      }
    }

To improve readability, this repository includes a script that improves LLDB's
understanding of Enoki types. With this script, the above turns into

    (enoki::Array<float, 3>) $0 = [1, 2, 3]